Moves hex functions to dicom-tag.cpp
