+++ include/dicom-element.h
@@ -2,12 +2,6 @@
 #include <core.h>
 #include <dicom-tag.h>
 
-#include <string>
-// convert decimal to hex, and pad the string to size `2*bytes`
-extern std::string DecToHex(uint64_t value, uint8_t bytes = 1);
-// convert up to 8 bytes of hex to decimal
-extern uint64_t HexToDec(std::string hex);
-
 /* class: DicomElement
  *  Takes an index and data buffer to a DICOM file.
  *  Using the buffer and index, it locates DICOM data element information.
