+++ include/dicom-tag.h
@@ -1,4 +1,10 @@
 #pragma once
+#include <string>
+// convert decimal to hex, and pad the string to size `2*bytes`
+extern std::string DecToHex(uint64_t value, uint8_t bytes = 1);
+// convert up to 8 bytes of hex to decimal
+extern uint64_t HexToDec(std::string hex);
+
 // Dicom Tags/VR/Name found here https://www.dicomlibrary.com/dicom/dicom-tags/
 enum dicomTag{
     Item = 0xe000fffe,
