+++ src/dicom-element.cpp
@@ -1,6 +1,5 @@
 #include <dicom-element.h>
 #include <dicom-tag.h>
-#include <sstream>
 
 uint32_t DicomElement::CalcLength() {
     // length width, and location must be deduced
@@ -35,23 +34,4 @@ uint64_t DicomElement::CalcSize() {
     return length + bytes;
 }
 
-std::string DecToHex(uint64_t value, uint8_t bytes) {
-    std::stringstream stream;
-    stream << std::hex << value;
-    std::string ret = std::string(stream.str());
-    if(ret.size() & 1){
-        ret.insert(0,"0");
-    }
-    // add 00 padding to ensure we reach the correct size
-    while(ret.size() < bytes*2){
-        ret.insert(0,"00");
-    }
-    return ret;
-}
-uint64_t HexToDec(std::string hex) {
-    uint64_t x;
-    std::stringstream ss;
-    ss << std::hex << hex;
-    ss >> x;
-    return x;
-}
+
